test: refactoring for naming convention
