+++ modules/identities/auxiliaries/authenticate/request_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/AssetMantle/modules/schema/helpers"
 	"github.com/AssetMantle/modules/schema/ids"
 	baseIDs "github.com/AssetMantle/modules/schema/ids/base"
-	"github.com/AssetMantle/modules/schema/lists/base"
-	base3 "github.com/AssetMantle/modules/schema/properties/base"
+	baseLists "github.com/AssetMantle/modules/schema/lists/base"
+	baseProperties "github.com/AssetMantle/modules/schema/properties/base"
 	baseQualified "github.com/AssetMantle/modules/schema/qualified/base"
 	"github.com/cosmos/cosmos-sdk/types"
 	"reflect"
@@ -17,9 +17,9 @@ import (
 )
 
 var (
-	mutableProperties      = base.NewPropertyList(base3.NewMetaProperty(baseIDs.NewStringID("authentication"), baseData.NewListData(base.NewDataList())))
-	immutableProperties    = base.NewPropertyList(base3.NewMetaProperty(baseIDs.NewStringID("ID1"), baseData.NewListData(base.NewDataList())))
-	toMutateMetaProperties = base.NewPropertyList(base3.NewMetaProperty(baseIDs.NewStringID("ID1"), baseData.NewListData(base.NewDataList(base.NewStringData("Test")))))
+	mutableProperties      = baseLists.NewPropertyList(baseProperties.NewMetaProperty(baseIDs.NewStringID("authentication"), baseData.NewListData(baseLists.NewDataList())))
+	immutableProperties    = baseLists.NewPropertyList(baseProperties.NewMetaProperty(baseIDs.NewStringID("ID1"), baseData.NewListData(baseLists.NewDataList())))
+	toMutateMetaProperties = baseLists.NewPropertyList(baseProperties.NewMetaProperty(baseIDs.NewStringID("ID1"), baseData.NewListData(baseLists.NewDataList(baseData.NewStringData("Test")))))
 	immutables             = baseQualified.NewImmutables(immutableProperties)
 	mutables               = baseQualified.NewMutables(mutableProperties)
 	testClassificationID   = baseIDs.NewClassificationID(immutables, mutables)
